sepsis-safe-offline-rl/requirements.txt
# Core RL and ML Libraries
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Offline RL
d3rlpy>=2.0.0

# Gym-Sepsis Environment
# gym-sepsis  # Install from source or pip if available

# Experiment Tracking
mlflow>=2.8.0
wandb>=0.15.0  # Alternative to MLflow

# Plotting and Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Causal Inference
dowhy>=0.9.0  # For causal analysis
econml>=0.14.0  # For confounding analysis

# Interpretability
shap>=0.42.0
lime>=0.2.0

# Statistical Analysis
statsmodels>=0.14.0
pingouin>=0.5.3  # For advanced statistics

# Uncertainty Quantification
uncertainty-toolbox>=0.1.0

# OPE (Off-Policy Evaluation)
# Note: Many OPE methods are custom implementations

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0  # For CLI

# Notebook Support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Type Checking
mypy>=1.5.0
types-PyYAML>=6.0.0
